Move project files out of storybook folder
