+++ src/components/SlidingUpPanelComponent.tsx
