+++ src/icons/SonrLogo.tsx
