+++ storybook/stories/PrimaryButton/PrimaryButton.stories.js
@@ -1,6 +1,6 @@
 import { storiesOf } from "@storybook/react-native";
 import React from "react";
-import PrimaryButton from ".";
+import PrimaryButton from "../../../src/components/PrimaryButton";
 import CenterViewDark from "../CenterViewDark";
 
 storiesOf("PrimaryButton", module)
