+++ storybook/stories/SecondaryButton/SecondaryButton.stories.js
@@ -1,6 +1,6 @@
 import { storiesOf } from "@storybook/react-native";
 import React from "react";
-import SecondaryButton from ".";
+import SecondaryButton from "../../../src/components/SecondaryButton";
 import CenterViewDark from "../CenterViewDark";
 
 storiesOf("SecondaryButton", module)
